feat:Product Update 기능 수정
상품 수정 버튼을 누르면 상품을 선택할 수 있고 선택한 상품을 한번에 수정한다.
InputDefault,LabelDefault, TextareaDefault.tsx
- fieldName이 true일 경우에 react-hook-form에서 데이터를 fieldName으로 묶을 수 있다.
productItemState.ts
- StateMenuBar.tsx를 제어하기 위해 selectOptionState atom을 추가하였다.
- option은 noen, delete, update이다.
AdminManageProductItemList, DeleteModalChildren.tsx
- setSelectState을 전역 상태로 관리하기 때문에 useState로 넘기던 값을 삭제하였다.
StateMenuBar.tsx
- setSelectState을 전역 상태로 관리하기 때문에 관련된 값을 추가하였다.
UpdateModalForm.tsx
- UpdateModalChildren.tsx에서 form필드를 생성하기 위한 컴포넌트이다.
UpdateModalChildren.tsx
- selectList로부터 값을 받아 사용자가 수정값을 입력하기 위한 컴포넌트이다.
- UpdateModalChildren.tsx의 값을 react-hook-form으로 받아 selectUpdateItemsSubmitHandler()에서 productList를 업데이트한다.
AdminMenu
- IsOpenModalChildren의 경로 변경하였다.
Router.tsx
- 크루 기능이 필요없어져서 route를 삭제하였다.
