+++ project/client/src/Components/Form/LabelDefault.tsx
@@ -1,15 +1,4 @@
 import React from "react";
-import styled from "styled-components";
-
-const LabelContainer = styled.div`
-  margin-top: 0.6rem;
-  align-self: baseline;
-  label {
-    font-size: 1.2rem;
-    text-align: left;
-    font-weight: 500;
-  }
-`;
 
 interface ILabelProps
   extends React.DetailedHTMLProps<
@@ -19,12 +8,12 @@ interface ILabelProps
   text: string | null | undefined;
 }
 
-const Label = ({ text, ...props }: ILabelProps) => {
+const LabelDefault = ({ text, ...props }: ILabelProps) => {
   return (
-    <LabelContainer>
+    <>
       <label {...props}>{text}</label>
-    </LabelContainer>
+    </>
   );
 };
 
-export default Label;
+export default LabelDefault;
