+++ project/client/src/state/productItemState.ts
@@ -1,5 +1,6 @@
 import { atom } from "recoil";
 import { productList, ProductListValues } from "../mockup/productList";
+import { SelectOption } from "../Page/Admin/AdminManageProductItemList";
 
 export const productListState = atom({
   key: "productList",
@@ -9,4 +10,9 @@ export const productListState = atom({
 export const selectProductListState = atom({
   key: "selectProductList",
   default: <ProductListValues[]>[],
+});
+
+export const selectOptionState = atom<SelectOption>({
+  key: "selectOption",
+  default: { option: "none" },
 });
